encapsulate whole atlas component on client
The imports of the atlas component were previously run on the server.
The deck and react-map-gl/maplibre imports cause issues directory import errors.
Moving the whole component to the client prevents these errors

closes #11
